api: merge crate nydus-app into crate nydus
Merge crate nydus-app into crate nydus, to reduce number of crates.
