--- app/CHANGELOG.md
@@ -1,14 +1,0 @@
-# Changelog
-## [Unreleased]
-
-### Added
-
-### Fixed
-
-### Deprecated 
-
-## [v0.1.0]
-
-### Added
-
-- Initial release
