--- app/CODEOWNERS
@@ -1,1 +1,0 @@
-* @bergwolf @imeoer @jiangliu
