--- app/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "nydus-app"
-version = "0.3.2"
-authors = ["The Nydus Developers"]
-description = "Application framework for Nydus Image Service"
-readme = "README.md"
-repository = "https://github.com/dragonflyoss/image-service"
-license = "Apache-2.0 OR BSD-3-Clause"
-edition = "2018"
-build = "build.rs"
-
-[build-dependencies]
-time = { version = "0.3.14", features = ["formatting"] }
-
-[dependencies]
-regex = "1.5.5"
-flexi_logger = { version = "0.25", features = ["compress"] }
-libc = "0.2"
-log = "0.4"
-nix = "0.24"
-serde = { version = "1.0.110", features = ["serde_derive"] }
-log-panics = { version = "2.1.0", features = ["with-backtrace"] }
-
-nydus-error = { version = "0.2", path = "../error" }
